% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{bits}
\alias{bits}
\title{Create a bits object}
\usage{
bits(x)
}
\arguments{
\item{x}{a logical vector}
}
\description{
Create a bits object
}
\seealso{
Other bits: 
\code{\link{as.character.bits}()},
\code{\link{bits2int}()},
\code{\link{c.bits}()},
\code{\link{int2bits}()},
\code{\link{print.bits}()}
}
\author{
Thierry Onkelinx
}
\concept{bits}
