% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{bits2int}
\alias{bits2int}
\title{Convert a bits object to an integer}
\usage{
bits2int(i)
}
\arguments{
\item{i}{the bits object}
}
\description{
Convert a bits object to an integer
}
\seealso{
Other bits: 
\code{\link{as.character.bits}()},
\code{\link{bits}()},
\code{\link{c.bits}()},
\code{\link{int2bits}()},
\code{\link{print.bits}()}
}
\author{
Thierry Onkelinx
}
\concept{bits}
