% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{c.bits}
\alias{c.bits}
\title{Combine bits}
\usage{
\method{c}{bits}(...)
}
\arguments{
\item{...}{the bits to concatenate}
}
\description{
The result inherits qrcode related arugments from the first element.
}
