% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Extract coordinates from a QR code object
Selects the dark elements from the \code{qr_code} object and returns their
coordinates.}
\usage{
coordinates(x)
}
\arguments{
\item{x}{the \code{qr_code} object.}
}
\value{
A \code{data.frame} with the \code{column} and \code{row} number of the dark
elements.
}
\description{
Extract coordinates from a QR code object
Selects the dark elements from the \code{qr_code} object and returns their
coordinates.
}
\examples{
x <- qr_code("test")
plot(x)
}
\seealso{
Other qr: 
\code{\link{generate_svg}()},
\code{\link{plot.qr_code}()},
\code{\link{print.qr_code}()},
\code{\link{qr_code}()},
\code{\link{qr_event}()},
\code{\link{qr_wifi}()}
}
\author{
Thierry Onkelinx
}
\concept{qr}
