% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataStringBinary.R
\name{DataStringBinary}
\alias{DataStringBinary}
\title{Function to convert input data string to binary polynomial}
\usage{
DataStringBinary(dataString, qrInfo)
}
\arguments{
\item{dataString}{Input data string.}

\item{qrInfo}{Dataframe that store all the required info to generate qrcode.}
}
\description{
Function to convert input data string to binary polynomial
}

