% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStringBinary.R, R/ECgenerator.R,
%   R/formatPolyGen.R, R/polynomialGenerator.R, R/qrFillUpMatrix.R,
%   R/qrInitMatrix.R, R/qrInterleave.R, R/qrMask.R, R/qrVersionInfo.R,
%   R/qrcode_gen.R, R/versionPolyGen.R
\name{DataStringBinary}
\alias{DataStringBinary}
\alias{ECgenerator}
\alias{formatPolyGen}
\alias{polynomialGenerator}
\alias{qrFillUpMatrix}
\alias{qrInitMatrix}
\alias{qrInterleave}
\alias{qrMask}
\alias{qrVersionInfo}
\alias{qrcode_gen}
\alias{versionPolyGen}
\title{Defunct legacy functions}
\usage{
DataStringBinary(...)

ECgenerator(...)

formatPolyGen(...)

polynomialGenerator(...)

qrFillUpMatrix(...)

qrInitMatrix(...)

qrInterleave(...)

qrMask(...)

qrVersionInfo(...)

qrcode_gen(...)

versionPolyGen(...)
}
\arguments{
\item{...}{obsolete arguments}
}
\description{
Defunct legacy functions
}
\author{
Victor Teh
}
\concept{legacy}
