% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ECgenerator.R
\name{ECgenerator}
\alias{ECgenerator}
\title{Error correction code generator
This function generate error correction code based on the input polynomial, Data polynomiail will be}
\usage{
ECgenerator(GenPoly, DataPoly, DCWordCount, ECWordCount)
}
\arguments{
\item{GenPoly}{Generated polynomial to calculate error correction code word}

\item{DataPoly}{Input data polynomial}

\item{DCWordCount}{Data code word count}

\item{ECWordCount}{Error code word count}
}
\value{
Error code word polynomial
}
\description{
Error correction code generator
This function generate error correction code based on the input polynomial, Data polynomiail will be
}

