% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPolyGen.R
\name{formatPolyGen}
\alias{formatPolyGen}
\title{Function to calculate and generate format polynomial}
\usage{
formatPolyGen(formatString, polyString)
}
\arguments{
\item{formatString}{QRcode format binary string}

\item{polyString}{polynomial to create ECL for formatString}
}
\description{
Function to calculate and generate format polynomial
}
\seealso{
Other legacy: 
\code{\link{DataStringBinary}()},
\code{\link{ECgenerator}()},
\code{\link{polynomialGenerator}()},
\code{\link{qrFillUpMatrix}()},
\code{\link{qrInitMatrix}()},
\code{\link{qrInterleave}()},
\code{\link{qrMask}()},
\code{\link{qrVersionInfo}()},
\code{\link{qrcode_gen}()},
\code{\link{versionPolyGen}()}
}
\author{
Victor Teh
}
\concept{legacy}
