% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_svg.R
\name{generate_svg}
\alias{generate_svg}
\title{Generate the QR code as an svg file}
\usage{
generate_svg(
  x,
  filename,
  size = 100,
  foreground = "black",
  background = "white",
  show = interactive(),
  ecl = c("L", "M", "H", "Q")
)
}
\arguments{
\item{x}{the input string}

\item{filename}{Where to store the filename.
Silently overwrites existings files.
Tries to create the path, when it doesn't exist.}

\item{size}{size of the svg file in pixels.}

\item{foreground}{Stroke and fill colour for the foreground.
Use a valid \href{https://www.w3schools.com/colors/}{CSS color}.
Defaults to \code{"black"}.}

\item{background}{Fill colour for the background.
Use a valid \href{https://www.w3schools.com/colors/}{CSS color}.
Defaults to \code{"white"}.}

\item{show}{Open the file after creating it.
Defaults to \code{TRUE} on \code{\link[=interactive]{interactive()}} sessions, otherwise \code{FALSE}.}

\item{ecl}{the required error correction level.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%).
Defaults to \code{"L"}.}
}
\value{
invisible \code{NULL}
}
\description{
Create the QR code using \code{\link[=qr_code]{qr_code()}} and save it as an svg file.
}
\examples{
generate_svg(
  x = "HELLO WORLD", filename = tempfile(fileext = ".svg"), show = FALSE
)
}
\seealso{
Other qr: 
\code{\link{plot.qr_code}()},
\code{\link{print.qr_code}()},
\code{\link{qr_code}()}
}
\author{
Thierry Onkelinx
}
\concept{qr}
