% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.qr_code}
\alias{plot.qr_code}
\title{Plot the QR code}
\usage{
\method{plot}{qr_code}(x, col = c("white", "black"), y, ...)
}
\arguments{
\item{x}{the \code{qr_code} object}

\item{col}{Define the colours.
The first element refers to \code{FALSE} and the second \code{TRUE}.
Defaults to \code{c("white", "black")}.}

\item{y}{currently ignored}

\item{...}{currently ignored}
}
\description{
Plot the QR code
}
\examples{
qr <- qr_code("HELLO WORLD")
plot(qr)
}
\seealso{
Other qr: 
\code{\link{generate_svg}()},
\code{\link{print.qr_code}()},
\code{\link{qr_code}()},
\code{\link{qr_event}()},
\code{\link{qr_wifi}()}
}
\author{
Thierry Onkelinx
}
\concept{qr}
