% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialGenerator.R
\name{polynomialGenerator}
\alias{polynomialGenerator}
\title{Function to generate polynomial}
\usage{
polynomialGenerator(ECcount)
}
\arguments{
\item{ECcount}{error correction code word count}
}
\value{
polynomail to generate Error correction code
}
\description{
Function to generate polynomial
}
\seealso{
Other legacy: 
\code{\link{DataStringBinary}()},
\code{\link{ECgenerator}()},
\code{\link{formatPolyGen}()},
\code{\link{qrFillUpMatrix}()},
\code{\link{qrInitMatrix}()},
\code{\link{qrInterleave}()},
\code{\link{qrMask}()},
\code{\link{qrVersionInfo}()},
\code{\link{qrcode_gen}()},
\code{\link{versionPolyGen}()}
}
\author{
Victor Teh
}
\concept{legacy}
