% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{print.bits}
\alias{print.bits}
\title{Print a bits vector}
\usage{
\method{print}{bits}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{currently ignored}
}
\description{
Print a bits vector
}
