% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_code.R
\name{qr_code}
\alias{qr_code}
\title{Generate the QR code}
\usage{
qr_code(x, ecl = c("L", "M", "Q", "H"))
}
\arguments{
\item{x}{the input string}

\item{ecl}{the required error correction level.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%)}
}
\value{
The QR code as a logical matrix with "qr_code" class.
}
\description{
Generate the QR code
}
\seealso{
Other qr: 
\code{\link{generate_svg}()},
\code{\link{plot.qr_code}()},
\code{\link{print.qr_code}()}
}
\author{
Thierry Onkelinx
}
\concept{qr}
