% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_error.R
\name{qr_error}
\alias{qr_error}
\title{Create the message and error code bit string}
\usage{
qr_error(x, ecl = c("L", "M", "Q", "H"))
}
\arguments{
\item{x}{the input string}

\item{ecl}{the required error correction level.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%).
Defaults to \code{"L"}.}
}
\value{
The message as a \code{bits()} object.
}
\description{
The function returns a bit string containing the message.
}
\examples{
qr_error("HELLO WORLD")
}
\seealso{
Other internal: 
\code{\link{qr_encode}()},
\code{\link{qr_matrix}()},
\code{\link{qr_mode}()},
\code{\link{qr_version}()}
}
\author{
Thierry Onkelinx
}
\concept{internal}
