% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_logo.R
\name{qr_logo}
\alias{qr_logo}
\title{Generate QR code with a logo on top}
\usage{
qr_logo(
  x,
  logo,
  ecl = c("L", "M", "Q", "H"),
  input_ecl = c("H", "Q", "M", "L")
)
}
\arguments{
\item{x}{the input string}

\item{logo}{the path to a logo image file}

\item{ecl}{the required error correction level for the remaining QR code
covered by the logo.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%).
Defaults to \code{"L"}.}

\item{input_ecl}{the required error correction level for the QR code
underneath the logo.
Must be higher than the selected level in \code{ecl}.
Defaults to \code{"H"}.
The difference between \code{ecl} and \code{input_ecl} determines the size of the logo.}
}
\description{
Will create an \code{svg} file with the QR code and logo.
}
