% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_matrix.R
\name{qr_matrix}
\alias{qr_matrix}
\title{Prepare matrices with default patterns and unmasked data}
\usage{
qr_matrix(x, ecl = c("L", "M", "Q", "H"))
}
\arguments{
\item{x}{the input string}

\item{ecl}{the required error correction level.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%)}
}
\description{
Prepare matrices with default patterns and unmasked data
}
\seealso{
Other internal: 
\code{\link{qr_encode}()},
\code{\link{qr_error}()},
\code{\link{qr_mode}()},
\code{\link{qr_version}()}
}
\author{
Thierry Onkelinx
}
\concept{internal}
