% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_mode.R
\name{qr_mode}
\alias{qr_mode}
\title{Determine the required mode}
\usage{
qr_mode(x)
}
\arguments{
\item{x}{the input string}
}
\value{
a character indicating the mode
}
\description{
Determine the required mode
}
\seealso{
Other internal: 
\code{\link{qr_encode}()},
\code{\link{qr_error}()},
\code{\link{qr_matrix}()},
\code{\link{qr_version}()}
}
\author{
Thierry Onkelinx
}
\concept{internal}
