% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_sepa.R
\name{qr_sepa}
\alias{qr_sepa}
\title{Generate a QR code for a SEPA payment}
\usage{
qr_sepa(
  iban,
  beneficiary,
  amount,
  unstructured_reference = "",
  bic = "",
  purpose = "",
  structured_reference = ""
)
}
\arguments{
\item{iban}{the IBAN of the beneficiary.}

\item{beneficiary}{the name of the beneficiary.}

\item{amount}{the amount to transfer. Must be in EUR.}

\item{unstructured_reference}{the unstructured reference.
The unstructured reference is a string of maximum 140 characters.}

\item{bic}{the BIC of the beneficiary.}

\item{purpose}{the purpose of the payment.}

\item{structured_reference}{the structured reference.}
}
\description{
Generate a QR code for a SEPA payment
}
