% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_version.R
\name{qr_version}
\alias{qr_version}
\title{Determine the required version}
\usage{
qr_version(x, ecl = c("L", "M", "Q", "H"))
}
\arguments{
\item{x}{the input string}

\item{ecl}{the required error correction level.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%)}
}
\description{
Determine the required version
}
