% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_wifi.R
\name{qr_wifi}
\alias{qr_wifi}
\title{Generate QR code with wifi login information}
\usage{
qr_wifi(
  ssid,
  encryption = c("WPA", "WEP", ""),
  key = "",
  hidden = FALSE,
  ecl = c("L", "M", "Q", "H")
)
}
\arguments{
\item{ssid}{The SSID of the network.}

\item{encryption}{The encryption standard.
Options are \code{"WPA"}, \code{"WEP"} and \code{""}.
The latter implies no encryption.
Defaults to \code{"WPA"}.}

\item{key}{The key for the encryption.}

\item{hidden}{Use \code{FALSE} for a visible SSID.
Use \code{TRUE} for a hidden SSID.
Defaults to  \code{FALSE}.}

\item{ecl}{the required error correction level.
Available options are \code{"L"} (7\%), \code{"M"} (15\%), \code{"Q"} (25\%) and \code{"H"} (30\%).
Defaults to \code{"L"}.}
}
\description{
Generate QR code with wifi login information
}
\seealso{
Other qr: 
\code{\link{coordinates}()},
\code{\link{generate_svg}()},
\code{\link{plot.qr_code}()},
\code{\link{print.qr_code}()},
\code{\link{qr_code}()},
\code{\link{qr_event}()}
}
\concept{qr}
