\name{qrcode-package}
\alias{qrcode-package}
\alias{qrcode}
\docType{package}
\title{
\packageTitle{qrcode}
}
\description{
\packageDescription{qrcode}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qrcode}
\packageIndices{qrcode}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{qrcode}

Maintainer: \packageMaintainer{qrcode}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
