% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qrFillUpMatrix.R
\name{qrFillUpMatrix}
\alias{qrFillUpMatrix}
\title{Function to fill up the data bits}
\usage{
qrFillUpMatrix(allBinary, data, version)
}
\arguments{
\item{allBinary}{all data in binary in character format.}

\item{data}{raster data created by \code{\link{qrFillUpMatrix}}}

\item{version}{version of the QRcode.}
}
\value{
raster data filled up with data bits
}
\description{
\code{qrFillUpMatrix} fill up the predefined QRcode raster.
}

