% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qrInitMatrix.R
\name{qrInitMatrix}
\alias{qrInitMatrix}
\title{Function to initialize qrcode in matrix for different version}
\usage{
qrInitMatrix(version)
}
\arguments{
\item{version}{version number of the target QRcode

\code{qrRaster} is used to create a basic structre of QRcode in raster format. Each element in QRcode will be marked as different value.}
}
\description{
Function to initialize qrcode in matrix for different version
}

