% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qrInterleave.R
\name{qrInterleave}
\alias{qrInterleave}
\title{Function to interleave the Data Code and Error Correction Core}
\usage{
qrInterleave(poly, dataPoly, qrInfo)
}
\arguments{
\item{poly}{Error Correction code word polynomial}

\item{dataPoly}{Input data code word polynomial}

\item{qrInfo}{Dataframe that store all the required info to
generate qrcode. Via \code{qrVersionInfo}}
}
\value{
Interleaved polynomial readied to fill up the qrcode matrix
}
\description{
Function to interleave the Data Code and Error Correction Core
}

