% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrVersionInfo.R
\name{qrVersionInfo}
\alias{qrVersionInfo}
\title{Function to identify the version of the QRcode based on input string}
\usage{
qrVersionInfo(dataString, ECLevel = c("L", "M", "Q", "H"))
}
\arguments{
\item{dataString}{dataString is the input string.}

\item{ECLevel}{Error Correction Level.
In QRcode standard, the are 4 levels "L", "M", "Q" and "H" which represent
7\%, 15\%, 20\% and 30\% data recovery capability.}
}
\value{
1 row dataframe that include all required info to generate QRcode.
}
\description{
Function to identify the version of the QRcode based on input string
}
