% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{bits}
\alias{bits}
\title{Create a bits object}
\usage{
bits(x)
}
\arguments{
\item{x}{a logical vector}
}
\description{
Converts a logical vector into a bits object.
This remains a logical vector.
The main difference is that is printed as a \code{0} and \code{1} bit string rather
than a \code{FALSE} and \code{TRUE} vector
}
\examples{
z <- bits(c(FALSE, TRUE))
z
str(z)
}
\seealso{
Other bits: 
\code{\link{as.character.bits}()},
\code{\link{bits2int}()},
\code{\link{c.bits}()},
\code{\link{print.bits}()}
}
\author{
Thierry Onkelinx
}
\concept{bits}
