% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{c.bits}
\alias{c.bits}
\title{Combine bits}
\usage{
\method{c}{bits}(...)
}
\arguments{
\item{...}{the bits to concatenate}
}
\description{
The result inherits qrcode related arugments from the first element.
}
\seealso{
Other bits: 
\code{\link{as.character.bits}()},
\code{\link{bits2int}()},
\code{\link{bits}()},
\code{\link{int2bits}()},
\code{\link{print.bits}()}
}
\author{
Thierry Onkelinx
}
\concept{bits}
