% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStringBinary.R
\name{DataStringBinary}
\alias{DataStringBinary}
\title{Function to convert input data string to binary polynomial}
\usage{
DataStringBinary(dataString, qrInfo)
}
\arguments{
\item{dataString}{input data string.}

\item{qrInfo}{dataframe that store all the required info to generate qrcode.}
}
\description{
Convert input data string to binary polynomial
}
