% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_string_binary.R, R/ec_generator.R,
%   R/format_poly_gen.R, R/polynomial_generator.R, R/qr_fill_up_matrix.R,
%   R/qr_init_matrix.R, R/qr_interleave.R, R/qr_mask.R, R/qr_version_info.R,
%   R/qrcode_gen.R, R/version_poly_gen.R
\name{DataStringBinary}
\alias{DataStringBinary}
\alias{ECgenerator}
\alias{formatPolyGen}
\alias{polynomialGenerator}
\alias{qrFillUpMatrix}
\alias{qrInitMatrix}
\alias{qrInterleave}
\alias{qrMask}
\alias{qrVersionInfo}
\alias{qrcode_gen}
\alias{versionPolyGen}
\title{Defunct legacy functions}
\usage{
DataStringBinary(...)

ECgenerator(...)

formatPolyGen(...)

polynomialGenerator(...)

qrFillUpMatrix(...)

qrInitMatrix(...)

qrInterleave(...)

qrMask(...)

qrVersionInfo(...)

qrcode_gen(...)

versionPolyGen(...)
}
\arguments{
\item{...}{obsolete arguments}
}
\description{
Defunct legacy functions
}
\author{
Victor Teh
}
\concept{legacy}
