% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECgenerator.R
\name{ECgenerator}
\alias{ECgenerator}
\title{Error correction code generator
Generate error correction code based on the input polynomial.}
\usage{
ECgenerator(GenPoly, DataPoly, DCWordCount, ECWordCount)
}
\arguments{
\item{GenPoly}{generated polynomial to calculate error correction code word}

\item{DataPoly}{input data polynomial}

\item{DCWordCount}{data code word count}

\item{ECWordCount}{error code word count}
}
\value{
Error code word polynomial
}
\description{
Error correction code generator
Generate error correction code based on the input polynomial.
}
\seealso{
Other legacy: 
\code{\link{DataStringBinary}()},
\code{\link{formatPolyGen}()},
\code{\link{polynomialGenerator}()},
\code{\link{qrFillUpMatrix}()},
\code{\link{qrInitMatrix}()},
\code{\link{qrInterleave}()},
\code{\link{qrMask}()},
\code{\link{qrVersionInfo}()},
\code{\link{qrcode_gen}()},
\code{\link{versionPolyGen}()}
}
\author{
Victor Teh
}
\concept{legacy}
