% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{int2bits}
\alias{int2bits}
\title{Convert an integer to a bits object}
\usage{
int2bits(i, n_bit = 16)
}
\arguments{
\item{i}{the integer}

\item{n_bit}{the number of bits}
}
\description{
Convert an integer to a bits object
}
\seealso{
Other bits: 
\code{\link{as.character.bits}()},
\code{\link{bits2int}()},
\code{\link{bits}()},
\code{\link{c.bits}()},
\code{\link{print.bits}()}
}
\author{
Thierry Onkelinx
}
\concept{bits}
