% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialGenerator.R
\name{polynomialGenerator}
\alias{polynomialGenerator}
\title{Function to generate polynomial}
\usage{
polynomialGenerator(ECcount)
}
\arguments{
\item{ECcount}{error correction code word count}
}
\value{
polynomail to generate Error correction code
}
\description{
Function to generate polynomial
}
