% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.qr_code}
\alias{print.qr_code}
\title{Print the qr_code object}
\usage{
\method{print}{qr_code}(x, ...)
}
\arguments{
\item{x}{the \code{qr_code} object}

\item{...}{currently ignored}
}
\description{
Please use \code{plot(x)} for a better quality image
}
\examples{
qr_code("HELLO WORLD")
}
\seealso{
Other qr: 
\code{\link{generate_svg}()},
\code{\link{plot.qr_code}()},
\code{\link{qr_code}()}
}
\author{
Thierry Onkelinx
}
\concept{qr}
