% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrInitMatrix.R
\name{qrInitMatrix}
\alias{qrInitMatrix}
\title{Function to initialize QRcode in matrix for different version}
\usage{
qrInitMatrix(version)
}
\arguments{
\item{version}{version number of the target QRcode}
}
\description{
Create a basic structre of QRcode in matrix format.
Each element in QRcode will be marked as different value.
}
