% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrInitMatrix.R
\name{qrInitMatrix}
\alias{qrInitMatrix}
\title{Function to initialize QRcode in matrix for different version}
\usage{
qrInitMatrix(version)
}
\arguments{
\item{version}{version number of the target QRcode}
}
\description{
Create a basic structre of QRcode in matrix format.
Each element in QRcode will be marked as different value.
}
\seealso{
Other legacy: 
\code{\link{DataStringBinary}()},
\code{\link{ECgenerator}()},
\code{\link{formatPolyGen}()},
\code{\link{polynomialGenerator}()},
\code{\link{qrFillUpMatrix}()},
\code{\link{qrInterleave}()},
\code{\link{qrMask}()},
\code{\link{qrVersionInfo}()},
\code{\link{qrcode_gen}()},
\code{\link{versionPolyGen}()}
}
\author{
Victor Teh
}
\concept{legacy}
