% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrInterleave.R
\name{qrInterleave}
\alias{qrInterleave}
\title{Function to interleave the Data Code and Error Correction Core}
\usage{
qrInterleave(poly, dataPoly, qrInfo)
}
\arguments{
\item{poly}{error correction code word polynomial}

\item{dataPoly}{input data code word polynomial}

\item{qrInfo}{dataframe that store all the required info to
generate QRcode. Via \code{qrVersionInfo}}
}
\value{
Interleaved polynomial readied to fill up the QRcode matrix
}
\description{
Function to interleave the Data Code and Error Correction Core
}
\seealso{
Other legacy: 
\code{\link{DataStringBinary}()},
\code{\link{ECgenerator}()},
\code{\link{formatPolyGen}()},
\code{\link{polynomialGenerator}()},
\code{\link{qrFillUpMatrix}()},
\code{\link{qrInitMatrix}()},
\code{\link{qrMask}()},
\code{\link{qrVersionInfo}()},
\code{\link{qrcode_gen}()},
\code{\link{versionPolyGen}()}
}
\author{
Victor Teh
}
\concept{legacy}
