% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionPolyGen.R
\name{versionPolyGen}
\alias{versionPolyGen}
\title{Function to calculate and generate version polynomial}
\usage{
versionPolyGen(versionString, polyString)
}
\arguments{
\item{versionString}{version in binary string}

\item{polyString}{polynomial in binary string, specified in the standard
to calculate version ECL.}
}
\value{
version polynomial.
}
\description{
Function to calculate and generate version polynomial
}
